chore: inline test utils
