+++ __tests__/AGENTS.md
@@ -1,7 +1,7 @@
 # AGENTS
 
 - Run tests with `npm test`. Required environment variables are listed in `README.md`.
-- `jest.setup.ts`, `jest.globalSetup.ts` and `jest.globalTeardown.ts` install an `undici.ProxyAgent` so `fetch` respects `HTTPS_PROXY`.
+- `jest.setup.ts`, `jest.globalSetup.ts` and `jest.globalTeardown.ts` import `setupProxyFetch` from `__tests__/helpers/` so `fetch` respects `HTTPS_PROXY`.
 - If a test fails with network errors, check that `HTTPS_PROXY` is set and that the `undici` dependency is installed.
 - Connectivity to `admin.googleapis.com` and other APIs is available once the proxy agent is active. A quick `node -e "require('undici'); fetch('https://admin.googleapis.com')"` should return a 404 status.
 - Live API endpoints may respond with 4xx errors if parameters do not match the Google Admin documentation. Review the endpoint helpers when this occurs.
