+++ __tests__/endpoints/ci/saml-profiles.test.ts
@@ -3,7 +3,7 @@ import {
   getIdpCreds,
   listSamlProfiles
 } from "@/app/lib/workflow/endpoints/ci";
-import { createLiveApiContext } from "../../../test-utils/live-api-context";
+import { createLiveApiContext } from "../../helpers/live-api-context";
 
 describe("SAML Profiles - Live API", () => {
   let apiContext: ReturnType<typeof createLiveApiContext>;
