+++ __tests__/endpoints/graph/applications.test.ts
@@ -1,5 +1,5 @@
 import { instantiateProv, instantiateSSO, listAppTemplates } from '@/app/lib/workflow/endpoints/graph';
-import { createLiveApiContext } from '../../../test-utils/live-api-context';
+import { createLiveApiContext } from '../../helpers/live-api-context';
 
 describe('Applications - Live API', () => {
   let apiContext: ReturnType<typeof createLiveApiContext>;
