+++ __tests__/endpoints/graph/policies.test.ts
@@ -1,5 +1,5 @@
 import { createPolicy, linkPolicy, listPolicies } from '@/app/lib/workflow/endpoints/graph';
-import { createLiveApiContext } from '../../../test-utils/live-api-context';
+import { createLiveApiContext } from '../../helpers/live-api-context';
 
 describe('Policies - Live API', () => {
   let apiContext: ReturnType<typeof createLiveApiContext>;
