+++ __tests__/endpoints/graph/saml-settings.test.ts
@@ -1,5 +1,5 @@
 import { getSamlSettings, patchSamlSettings } from '@/app/lib/workflow/endpoints/graph';
-import { createLiveApiContext } from '../../../test-utils/live-api-context';
+import { createLiveApiContext } from '../../helpers/live-api-context';
 
 describe('SAML Settings - Live API', () => {
   let apiContext: ReturnType<typeof createLiveApiContext>;
