+++ __tests__/endpoints/graph/sync.test.ts
@@ -1,5 +1,5 @@
 import { getSync, patchSync, startSyncJob } from '@/app/lib/workflow/endpoints/graph';
-import { createLiveApiContext } from '../../../test-utils/live-api-context';
+import { createLiveApiContext } from '../../helpers/live-api-context';
 
 describe('Sync - Live API', () => {
   let apiContext: ReturnType<typeof createLiveApiContext>;
