+++ jest.setup.ts
@@ -1,19 +1,8 @@
 // Global test setup and teardown
-import { ProxyAgent, setGlobalDispatcher } from "undici";
+import { setupProxyFetch } from "./__tests__/helpers/proxyFetch";
 
 // Configure fetch to respect HTTPS proxy if set.
-const proxy = process.env.https_proxy || process.env.HTTPS_PROXY;
-if (proxy) {
-  const agent = new ProxyAgent(proxy);
-  setGlobalDispatcher(agent);
-  const origFetch: typeof fetch = globalThis.fetch;
-  type FetchInitWithDispatcher = RequestInit & { dispatcher: typeof agent };
-
-  globalThis.fetch = ((
-    input: RequestInfo | URL,
-    init?: RequestInit,
-  ) => origFetch(input, { ...init, dispatcher: agent } as FetchInitWithDispatcher)) as typeof fetch;
-}
+setupProxyFetch();
 
 beforeAll(async () => {
   // Validate test environment
