+++ tsconfig.spec.json
@@ -5,7 +5,6 @@
     "**/*.test.ts",
     "**/*.test.tsx",
     "__tests__/**/*.ts",
-    "test-utils/**/*.ts",
     "jest.config.ts",
     "jest.globalSetup.ts",
     "jest.globalTeardown.ts"
